requirements.txt
app~=0.0.1
Flask~=3.1.0
Flask-Migrate~=4.1.0
Werkzeug~=3.1.3
itsdangerous~=2.2.0
SQLAlchemy~=2.0.40
alembic~=1.15.2
PyJWT~=2.10.1
flask-swagger-ui~=4.11.1
python-dotenv~=1.1.0
db~=0.1.1
